Add a recombiner program.
The recombiner program accepts input produced by a run of the program created
by wt_concurrency. The recombiner program then outputs an ascii art table of the
results from the sequence table. This table resembles the input to
wt_concurrency, but includes the results from reads and any WiredTiger errors.